Use React Tippy to render the dropdown and tooltips
