+++ package.json
@@ -32,7 +32,6 @@
     "babel-plugin-import": "^1.13.6",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-tippy": "^1.4.0",
     "react-virtuoso": "^4.6.2"
   },
   "devDependencies": {
